requirements.txt
certifi==2020.12.5
chardet==4.0.0
click==7.1.2
cycler==0.10.0
Flask==1.1.2
idna==2.10
itsdangerous==1.1.0
Jinja2==2.11.3
joblib==1.0.1
jsonpickle==2.0.0
kiwisolver==1.3.1
MarkupSafe==1.1.1
matplotlib==3.4.1
np==1.0.2
numpy==1.20.2
opencv-python==4.5.1.48
Pillow==8.2.0
playsound==1.2.2
pydub==0.25.1
pyparsing==2.4.7
python-dateutil==2.8.1
requests==2.25.1
simpleaudio==1.0.4
six==1.15.0
threadpoolctl==2.1.0
torch==1.8.1
torchvision==0.9.1
typing-extensions==3.7.4.3
urllib3==1.26.4
Werkzeug==1.0.1
